+++ mail/mail-tui/src/keychain.rs
@@ -1,44 +1,49 @@
 use crate::app_model::APP_ID;
 use anyhow::anyhow;
+use proton_core_common::os::{KeyChain, KeyChainEntryKind, KeyChainError, KeyChainExt};
 use proton_mail_common::proton_core_common::db::account::SessionEncryptionKey;
-use proton_mail_common::proton_core_common::os::{KeyChain, KeyChainError};
 use secrecy::{ExposeSecret, SecretString};
 use std::error::Error;
 use std::sync::Arc;
 
 pub struct AppKeyChain {
-    entry: Arc<keyring::Entry>,
+    session_key: Arc<keyring::Entry>,
 }
 
 impl AppKeyChain {
     pub fn new() -> Result<Self, Box<dyn Error>> {
         let entry = keyring::Entry::new(APP_ID, "session_key")?;
         Ok(Self {
-            entry: Arc::new(entry),
+            session_key: Arc::new(entry),
         })
     }
 
     pub fn init(&mut self) -> Result<(), Box<dyn Error>> {
-        let v = self.get()?;
+        let v = self.load::<SessionEncryptionKey>()?;
         if v.is_none() {
             let key = SessionEncryptionKey::random();
-            self.store(key.to_base64())?;
+            self.store(key)?;
         }
         Ok(())
+    }
+
+    fn kind_to_entry(&self, kind: KeyChainEntryKind) -> &Arc<keyring::Entry> {
+        match kind {
+            KeyChainEntryKind::EncryptionKey => &self.session_key,
+        }
     }
 }
 
 impl KeyChain for AppKeyChain {
-    fn store(&self, key: String) -> Result<(), KeyChainError> {
-        let key = SecretString::new(key);
-        self.entry
+    fn store_entry(&self, kind: KeyChainEntryKind, key: SecretString) -> Result<(), KeyChainError> {
+        self.kind_to_entry(kind)
             .set_password(key.expose_secret())
             .map_err(|e| KeyChainError::new(anyhow!(e).into()))?;
         Ok(())
     }
 
-    fn delete(&self) -> Result<(), KeyChainError> {
-        if let Err(e) = self.entry.delete_credential() {
+    fn delete_entry(&self, kind: KeyChainEntryKind) -> Result<(), KeyChainError> {
+        if let Err(e) = self.kind_to_entry(kind).delete_credential() {
             if !matches!(e, keyring::Error::NoEntry) {
                 return Err(KeyChainError::new(anyhow!(e).into()));
             }
@@ -46,9 +51,9 @@ impl KeyChain for AppKeyChain {
         Ok(())
     }
 
-    fn get(&self) -> Result<Option<String>, KeyChainError> {
-        match self.entry.get_password() {
-            Ok(str) => Ok(Some(str)),
+    fn load_entry(&self, kind: KeyChainEntryKind) -> Result<Option<SecretString>, KeyChainError> {
+        match self.kind_to_entry(kind).get_password() {
+            Ok(str) => Ok(Some(SecretString::new(str))),
             Err(e) => match e {
                 keyring::Error::NoEntry => Ok(None),
                 _ => Err(KeyChainError::new(anyhow!(e).into())),